Working on profile screen. Extra space issue is resolved by using lazyVstack.
