+++ TWB-Ecommerce/Base Components/Constants/Color+Constants.swift
@@ -15,4 +15,5 @@ struct Constants {
     static let black: Color = .init(hex: "#000000")
     static let white: Color = .init(hex: "#FFFFFF")
     static let clear: Color = .clear
+    static let grayBorder: Color = .init(hex: "#D9D9D9")
 }
