--- TWB-Ecommerce/Features/Profile/LoginProfileView.swift
@@ -1,18 +1,0 @@
-//
-//  LoginProfileView.swift
-//  TWB-Ecommerce
-//
-//  Created by Hassam Ali on 24/09/2024.
-//
-
-import SwiftUI
-
-struct LoginProfileView: View, LoginSignUpViews {
-    var body: some View {
-        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
-    }
-}
-
-#Preview {
-    ProfileView()
-}
